fix: categoryId 자료형 변경(int -> Long)
API 명세 오류로 인한 카테고리 아이디 자료형 정정

+++ server/src/main/java/team/youngcha/domain/option/dto/DefaultOptionSummary.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 @Getter
 public class DefaultOptionSummary {
     private final String name;
-    private final int categoryId;
+    private final Long categoryId;
     private final String imgUrl;
 
-    public DefaultOptionSummary(String name, int categoryId, String imgUrl) {
+    public DefaultOptionSummary(String name, Long categoryId, String imgUrl) {
         this.name = name;
         this.categoryId = categoryId;
         this.imgUrl = imgUrl;
